pyproject.toml
[project]
name = "styletts2-lite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.6.0",
    "einops>=0.8.1",
    "einops-exts>=0.0.4",
    "librosa>=0.11.0",
    "matplotlib>=3.10.3",
    "monotonic-align",
    "munch>=4.0.0",
    "nltk>=3.9.1",
    "noisereduce>=3.0.3",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "phonemizer-fork>=3.3.2",
    "pydub>=0.25.1",
    "pyyaml>=6.0.2",
    "soundfile>=0.13.1",
    "tensorboard>=2.19.0",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "typing>=3.10.0.0",
    "typing-extensions>=4.13.2",
]

[tool.uv.sources]
monotonic-align = { git = "https://github.com/resemble-ai/monotonic_align.git" }
